requirements.txt
pandas
streamlit
sentence-transformers
faiss-cpu
scikit-learn
